Learning with Figaro
==> Fixed the learning on a seq of team.
