+++ app/component/soccer/TeamHelper.scala
@@ -4,7 +4,7 @@ import component.Model
 
 class TeamHelper {
 
-  def observeEvidence(model: Model, hasHGA: Option[Boolean], hasForm: Option[Boolean]) {
+  def observeEvidence(model: Model, hasHGA: Option[Boolean], hasForm: Option[Boolean], hasGoodRating: Option[Boolean]) {
     hasForm match {
       case Some(value) => model.hasGoodForm.observe(value)
       case None => ()
@@ -12,6 +12,11 @@ class TeamHelper {
 
     hasHGA match {
       case Some(value) => model.hasHomeGroundAdvantage.observe(value)
+      case None => ()
+    }
+
+    hasGoodRating match {
+      case Some(value) => model.hasGoodRating.observe(value)
       case None => ()
     }
   }
