+++ conf/application.conf
@@ -61,7 +61,7 @@ cassandra {
 
   #host = ["197.189.214.210"]
   #host =["155.238.208.150"]
-  host =["155.238.215.122"]
+  host =["155.238.217.152"]
   #host =["192.168.140.42"]
   masterKeyspace = "k_master"
   pseudomasterKeyspace = "k_pseudomaster"
